Avoid attempt at purge of already purged session
The purge method can be called from both the disconnect message handler
and from the connectionLost handler. If the disconnect happened normally
then the connectionLost handler doesn't need to purge any more.

+++ broker/src/main/java/io/moquette/broker/SessionRegistry.java
@@ -264,7 +264,7 @@ public class SessionRegistry {
 
     void connectionClosed(Session session) {
         session.disconnect();
-        if (session.expireImmediately()) {
+        if (session.expireImmediately()&& !session.hasState(SessionStatus.DESTROYED)) {
             purgeSessionState(session);
             return;
         } else {
